add --home flags
After providing the ability to specify the location of the container's read-write layer, you can use the server's multiple disks to improve IO capabilities, or the read-write layer is placed on the remote disk, and the container is migrated from one host to another. The read-write layer corresponding to the old container can be reused.

Signed-off-by: yifang.jy <yifang.jy@alibaba-inc.com>