feature: support multiple snapshotter
Pouch support different runtimes, like runc, kata. For kata, default p9
filesystem has ready slow io, so we do some enhance for the file system
kata use(qcow2 block device). This introduce a big problem here, for
runc, the default snapshotter is overlay, for kata, we use qcow2 for
enhance io, but runc can not use qcow2, since it is a block device, kata
can not use overlay if it want a better performance in io.
So we want a strategy to make two runtimes happy.

We would use hook plugin to implement different snapshotter, so for now,
different snapshotter is not provided through pouch by default.
Snapshotter is set through context, the context can not set by user.

Signed-off-by: Ace-Tang <aceapril@126.com>