CRI: Add daemon configuration to StatusResponse as extra information
StatusResponse in CRI returns runtime extra information which is in
the map of Info. Now there is only one key in this map. This patch
add "daemon-config" as another key and its value is the daemon
configuration in json format. With this patch, for example, we can
get pouch root dir from StatusResponse.

Signed-off-by: Wang Rui <baijia.wr@antfin.com>