test: refact travis-ci
refact travis-ci, make it run faster.

1. remove CRI v1alpha1 test
2. split the integration into multiple parts, let them run together,
because the limit of travis ci jobs, we only can split four parts.
3. update travis-ci os version to ubuntu-16.04 xenial,
because in this version, we can use apt-get to install criu, instead of
building it by ourself, it can reduce the setup time.
4. I have built `nsenter` and upload to aliyun oss storage server,
so only download it instead of building it every time.

Signed-off-by: Rudy Zhang <rudyflyzhang@gmail.com>