test: make test run happy
test `TestNetworkDefault`:
In host, it get the default route, so it should get default route in
container as same. Some server has mtu information in host, so we should
remove it.

test `TestRunWithBridgeNetwork`:
some server has configed the dns: 127.0.0.1, but it has been removed in
container by libnetwork, it causes the different from host dns config,
so we should remove dns: 127.0.0.1 when we compare them.

test `TestRunWithoutCapability`:
change `chown` command.

Signed-off-by: Rudy Zhang <rudyflyzhang@gmail.com>