daemon/mgr/image: update deadline for loading images
PouchContainer will load all the images from containerd into memory. At
the beginning, we assume that it can load it in 10 secs. But if the
system has busy IO, it will take long time to load it, especially the
more-layers and huge-size images. So update it from 10 secs to 10 mins.

Signed-off-by: Wei Fu <fuweid89@gmail.com>