cli/daemon/apis/test: support to load docker images without name
For the docker image format, the tar file will contain the image
reference information. For this case, when we load the docker images, we
should not set the image name. If there are too many images and you just
want to load one of them, you can set the name.

Signed-off-by: Wei Fu <fuweid89@gmail.com>